pyproject.toml
[tool.poetry]
name = "bulk-update-tests"
version = "0.1.0"
description = ""
authors = ["Mikulas Poul <mikulaspoul@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "~=2.2.9"
django-bulk-update = "^2.2.0"
psycopg2-binary = "^2.8.4"
factory_boy = "^2.12.0"
ipython = "^7.11.1"
django-extensions = "^2.2.6"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | venv
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
